# Expand precompilation