--- a/.github/workflows/system.yml
+++ b/.github/workflows/system.yml
@@ -20,7 +20,6 @@ jobs:
         architecture: [x64]
         python-version: ['3.x']
         julia-version:
-          - '1.0'
           - '1'
           - 'nightly'
         include: