--- a/Project.toml
+++ b/Project.toml
@@ -16,7 +16,7 @@ VersionParsing = "81def892-9a0e-5fdd-b105-ffc91e053289"
 Conda = "1.0"
 MacroTools = "0.4, 0.5"
 VersionParsing = "1.0"
-julia = "0.7, 1.0"
+julia = "1.4"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"