This adds more extensive precompilation.

Eventually it would be better to switch to SnoopCompile.jl, but for some reason I kept running into segfaults with it when trying to snoop compile some example code. I think using SnoopCompile.jl with calls to the Python interpreter triggers some unexpected behavior, and will take some fixing before it works.

So in the meantime, this is something to improve the results of precompilation. As noted in https://github.com/MilesCranmer/PySR/pull/230, it seems like PyCall.jl is a source of slow startup for Python-Julia sysimages, so these additional precompile statements should help quite a bit.

It was generated with `--trace-compile=stderr` on a run of PySR and then manually parsed down and sorted. Thus it does not hit all functions by any means, but the precompilation statements are general and should help any PyCall.jl package.

Let me know what you think.
Cheers,
Miles