@MilesCranmer 2022-12-02
It's roughly a 30% improvement in startup time on this quick benchmark which does some simple calls. Code and timings below.

- Timing before: 
    - First run: 0.704 s
    - Second run: 0.063 s
    - (First - second): 0.641 s
- Timing after:
    - First run: 0.512 s
    - Second run: 0.063 s
    - (First - second): 0.449 s

This is the benchmarked code (in Python)
```python
from julia import Main as jl
from julia import LinearAlgebra

# Arrays:
x = jl.randn(jl.Float32, 10_000)
y = jl.randn(jl.Float32, 10_000)

# Functions:
LinearAlgebra.dot(x, y)

# String:
jl.s = "test"

# Tuple:
jl.t = (1, 2, 3)

# Dict:
jl.d = {"a": 1, "b": 2}
```

<details><summary>Code:</summary>

Run with: `julia bench_pycall.jl 20`

`bench_pycall.py`

```python
from julia import Main as jl
from julia import LinearAlgebra
import time

def f():

    # Arrays:
    x = jl.randn(jl.Float32, 10_000)
    y = jl.randn(jl.Float32, 10_000)

    # Functions:
    LinearAlgebra.dot(x, y)

    # String:
    jl.s = "test"

    # Tuple:
    jl.t = (1, 2, 3)

    # Dict:
    jl.d = {"a": 1, "b": 2}


start = time.time()
f()
end = time.time()
print(end - start, end=",")

start = time.time()
f()
end = time.time()
print(end - start, end="")
```

`bench_pycall.jl`

```julia
using Statistics

NTRIALS = parse(Int, ARGS[1])

function run_bench()
    out = read(`python bench_pycall.py`, String)
    out = split(out, ',')
    return [parse(Float64, x) for x in out]
end

recordings = []

for i in 1:NTRIALS
    push!(recordings, run_bench())
    println("Trial $i done, with $(recordings[end])")
end

first_run = median(rec[1] for rec in recordings)
second_run = median(rec[2] for rec in recordings)

print("Median first: $(first_run) s, median second: $(second_run) s")
```

</details>