@MilesCranmer 2022-12-02
I just added more precompilation calls to this PR using PyCall.jl's test suite. I now get a 38% speedup at that startup test. (now a 0.469 s first run, 0.071 s second run).

To get the extra precompilation calls, I replaced `@test` and `@testset` with identity macros, then ran `test/runtests.jl` manually with `--trace-compile=stderr`. I then manually sorted the precompile statements, and removed calls which were overly-specific to the tests, such as those belonging to specific library calls.

---

By the way, I thought a bit more about how we could use SnoopCompile.jl instead, and think it will be difficult, because one needs to call `Pkg.build("PyCall")` first before actually running the examples... So having a manual list of precompile statements like this might actually be the only option.