@stevengj 2022-12-03
Looks like `Core.gTypeofVararg` requires a more recent Julia version, i.e. one should update the requirements in Project.toml and CI.