@MilesCranmer 2022-12-03
Thanks, I removed those lines and now it's working on Julia >=1.4. I will set that as the Project.toml bound to be safe.

Implemented the CI changes too. I'm not sure what the other errors are from - any idea?