@MilesCranmer 2022-12-03
Okay it looks like the other errors are unrelated to this PR - venv test broken on master too: https://github.com/MilesCranmer/PyCall.jl/actions/runs/3606216417/jobs/6077346109.

Let me know if the PR is good to go.