@MilesCranmer 2023-01-04
@stevengj do you think this could be merged? Would be awesome to take full advantage of Julia 1.9 startup times on PyJulia-based apps.