@MilesCranmer 2023-01-05
I think auto-merge doesn't work when the CI is changed (Julia 1.0 was removed from the testing). The rest look like they pass though.