@MilesCranmer 2023-01-05
I think when a PR changes the CI settings (here, removing one of the Julia versions), then the other jobs are left as "waiting to complete." I've seen this on other PRs too. I think it is simply to prevent auto-merging when CI is accidentally turned off.